+++ include/render/Integrator.h
@@ -124,6 +124,7 @@ namespace render {
     bool intersectionBackendPrefersAnyHitBatch{false};
     bool intersectionBackendSupportsResidentFrontiers{false};
     bool intersectionBackendSupportsGpuFrontierCompaction{false};
+    bool intersectionBackendSupportsPreparedRayBatchCompaction{false};
     bool intersectionBackendSupportsResidentDirectLightBatches{false};
     double intersectionWorkerSeconds{0.0};
     double shadingWorkerSeconds{0.0};